Refactor the logging.
Started a refactor of the logging mechanism used.  Looking to simplify
the process in the c# code by using one logger - Log4Net.  If it is
deemed necessary to have 'multiple loggers' logging to n locations then
this can be done by leveraging the capabilities of Log4Net rather than
injecting n loggers into the code and then iterating over them in the
codebase.